sửa thanh máu khi nhân vật 100 máu, sửa dame và máu của quái, sửa tính toán máu của quái và nhân vật

+++ Assets/Scripts/GamePlay/Healingth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Healingth : MonoBehaviour
 {
-    public float healAmount;
+    public int healAmount;
 
     private void OnTriggerEnter2D(Collider2D other)
     {
